arch-x86,cpu: Override the int div latency local to x86.
Remove the ISA check when selecting the default integer division latency
for O3. Instead, create a parameter which the latency defaults to using
a proxy, and then override that parameter when setting the default
function unit pool in the x86 version of that CPU. The value can still
be overridden if the user wants, but this way it gets its specialized
value within x86 with no ISA check.

Change-Id: I1ef9ee94f4b16aebe03e043df5cdc6167efe6e64